+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - compressors are made optional components
 - fixed bug in `calc_connected_components`
+- corrected models of resistor physics
 
 ## v0.8